feat: rename http and place it in the good place
